+++ Navigation/Models/PostVK.swift
@@ -1,0 +1,18 @@
+//
+//  PostVK.swift
+//  Navigation
+//
+//  Created by Kirill Komov on 29.07.2021.
+//  Copyright © 2021 Artem Novichkov. All rights reserved.
+//
+import UIKit
+import iOSIntPackage
+
+struct PostVK {
+    let author: String
+    let description: String
+    let image: String
+    let filter: ColorFilter
+    let likes: Int
+    let views: Int
+}
