--- Navigation/MyCustomButton.swift
@@ -1,37 +1,0 @@
-//
-//  CustomButton.swift
-//  Navigation
-//
-//  Created by Kirill Komov on 12.01.2022.
-//  Copyright © 2022 Artem Novichkov. All rights reserved.
-//
-
-import UIKit
-
-class MyCustomButton: UIButton {
-    
-    private let buttonAction: () -> Void
-    
-    init(title: String?, titleColor: UIColor?, backgroundColor: UIColor?, backgroundImage: UIImage?, buttonAction: @escaping () -> Void) {
-        self.buttonAction = buttonAction
-        super.init(frame: .zero)
-        
-        self.setTitle(title, for: .normal)
-        self.setTitleColor(titleColor, for: .normal)
-        self.backgroundColor = backgroundColor
-        self.setBackgroundImage(backgroundImage, for: .normal)
-        self.translatesAutoresizingMaskIntoConstraints = false
-        
-        self.addTarget(self, action: #selector(buttonIsTapped), for: .touchUpInside)
-    }
-    
-    @objc private func buttonIsTapped() {
-        self.buttonAction()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    
-}
