+++ Navigation/ProfileUser/Photogallery/PhotosViewController.swift
@@ -1,0 +1,107 @@
+//
+//  PhotosViewController.swift
+//  Navigation
+//
+//  Created by Kirill Komov on 29.07.2021.
+//  Copyright © 2021 Artem Novichkov. All rights reserved.
+//
+
+import UIKit
+import iOSIntPackage
+
+class PhotosViewController: UIViewController {
+    
+    private let facade = ImagePublisherFacade()
+    private let layout = UICollectionViewFlowLayout()
+    private lazy var photoCollectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
+    private let collectionCellID = "collectionCellID"
+    private let arrayOfPhotos = PhotoStorage.photoArray
+    private var arrayOfPublishedPhotos: [UIImage] = []
+    
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        navigationController?.navigationBar.isHidden = false
+        navigationItem.title = "Photo Gallery"
+        navigationController?.navigationBar.topItem?.title = "Back"
+        
+        setupCollectionView()
+        photoCollectionView.backgroundColor = .white
+        
+        facade.addImagesWithTimer(time: 0.5, repeat: 21, userImages: arrayOfPhotos)
+    }
+    
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+        facade.subscribe(self)
+        
+    }
+    
+    override func viewWillDisappear(_ animated: Bool) {
+        super.viewWillDisappear(animated)
+        
+        navigationController?.navigationBar.isHidden = true
+        facade.removeSubscription(for: self)
+        
+    }
+    
+    private func setupCollectionView(){
+        view.addSubview(photoCollectionView)
+        photoCollectionView.translatesAutoresizingMaskIntoConstraints = false
+        photoCollectionView.dataSource = self
+        photoCollectionView.delegate = self
+        photoCollectionView.register(PhotosCollectionViewCell.self, forCellWithReuseIdentifier: "collectionCellID")
+        
+        let constraints = [
+            photoCollectionView.topAnchor.constraint(equalTo: view.topAnchor),
+            photoCollectionView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
+            photoCollectionView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
+            photoCollectionView.trailingAnchor.constraint(equalTo: view.trailingAnchor)
+        ]
+        
+        NSLayoutConstraint.activate(constraints)
+    }
+}
+
+extension PhotosViewController: UICollectionViewDataSource {
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return arrayOfPublishedPhotos.count
+    }
+
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        let cell: PhotosCollectionViewCell = collectionView.dequeueReusableCell(withReuseIdentifier: "collectionCellID", for: indexPath) as! PhotosCollectionViewCell
+        
+        cell.photo = arrayOfPublishedPhotos[indexPath.item]
+        return cell
+    }
+}
+
+extension PhotosViewController: UICollectionViewDelegateFlowLayout {
+    
+    private var baseInset: CGFloat { return 8 }
+    
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+        return CGSize(width: (photoCollectionView.frame.width - baseInset * 4) / 3, height: (photoCollectionView.frame.width - baseInset * 4) / 3)
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumLineSpacingForSectionAt section: Int) -> CGFloat {
+        return baseInset
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumInteritemSpacingForSectionAt section: Int) -> CGFloat {
+        return .zero
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, insetForSectionAt section: Int) -> UIEdgeInsets {
+        return UIEdgeInsets(top: baseInset, left: baseInset, bottom: .zero, right: baseInset)
+    }
+}
+
+extension PhotosViewController: ImageLibrarySubscriber {
+    func receive(images: [UIImage]) {
+        arrayOfPublishedPhotos = images        
+        let indexPath = IndexPath(item: images.count - 1, section: 0)
+        photoCollectionView.insertItems(at: [indexPath])
+    }
+}
