--- Navigation/Storage/PhotoStorage.swift
@@ -1,34 +1,0 @@
-//
-//  PhotoStorage.swift
-//  Navigation
-//
-//  Created by Kirill Komov on 29.07.2021.
-//  Copyright © 2021 Artem Novichkov. All rights reserved.
-//
-
-import UIKit
-
-struct PhotoStorage {
-    static let photoArray = [
-        #imageLiteral(resourceName: "02"),
-        #imageLiteral(resourceName: "01"),
-        #imageLiteral(resourceName: "16"),
-        #imageLiteral(resourceName: "11"),
-        #imageLiteral(resourceName: "06"),
-        #imageLiteral(resourceName: "19"),
-        #imageLiteral(resourceName: "05"),
-        #imageLiteral(resourceName: "20"),
-        #imageLiteral(resourceName: "10"),
-        #imageLiteral(resourceName: "15"),
-        #imageLiteral(resourceName: "17"),
-        #imageLiteral(resourceName: "14"),
-        #imageLiteral(resourceName: "18"),
-        #imageLiteral(resourceName: "04"),
-        #imageLiteral(resourceName: "13"),
-        #imageLiteral(resourceName: "12"),
-        #imageLiteral(resourceName: "08"),
-        #imageLiteral(resourceName: "03"),
-        #imageLiteral(resourceName: "07"),
-        #imageLiteral(resourceName: "09")]
-}
-
