+++ Navigation/Table/TableViewModelType.swift
@@ -1,0 +1,14 @@
+//
+//  TableViewModelType.swift
+//  Navigation
+//
+//  Created by Kirill Komov on 15.01.2022.
+//  Copyright © 2021 Artem Novichkov. All rights reserved.
+//
+
+import Foundation
+
+protocol TableViewModelType {
+    var numberOfRows: Int { get }
+    var postArray: [PostVK] { get }
+}
